Fix multigoal metaworld
