+++ data/envs/metaworld/generate_dataset.py
@@ -1,7 +1,7 @@
 import os
 from typing import Any, Dict, List, Optional
 
-import gym
+import gymnasium as gym
 import metaworld  # noqa: F401
 import numpy as np
 import torch
@@ -20,60 +20,6 @@ from sample_factory.utils.typing import Config
 from tqdm import tqdm
 
 
-ENV_NAMES = [
-    "assembly-v2",
-    "basketball-v2",
-    "bin-picking-v2",
-    "box-close-v2",
-    "button-press-topdown-v2",
-    "button-press-topdown-wall-v2",
-    "button-press-v2",
-    "button-press-wall-v2",
-    "coffee-button-v2",
-    "coffee-pull-v2",
-    "coffee-push-v2",
-    "dial-turn-v2",
-    "disassemble-v2",
-    "door-close-v2",
-    "door-lock-v2",
-    "door-open-v2",
-    "door-unlock-v2",
-    "drawer-close-v2",
-    "drawer-open-v2",
-    "faucet-close-v2",
-    "faucet-open-v2",
-    "hammer-v2",
-    "hand-insert-v2",
-    "handle-press-side-v2",
-    "handle-press-v2",
-    "handle-pull-side-v2",
-    "handle-pull-v2",
-    "lever-pull-v2",
-    "peg-insert-side-v2",
-    "peg-unplug-side-v2",
-    "pick-out-of-hole-v2",
-    "pick-place-v2",
-    "pick-place-wall-v2",
-    "plate-slide-back-side-v2",
-    "plate-slide-back-v2",
-    "plate-slide-side-v2",
-    "plate-slide-v2",
-    "push-back-v2",
-    "push-v2",
-    "push-wall-v2",
-    "reach-v2",
-    "reach-wall-v2",
-    "shelf-place-v2",
-    "soccer-v2",
-    "stick-pull-v2",
-    "stick-push-v2",
-    "sweep-into-v2",
-    "sweep-v2",
-    "window-close-v2",
-    "window-open-v2",
-]
-
-
 def make_custom_env(
     full_env_name: str,
     cfg: Optional[Dict] = None,
@@ -84,7 +30,7 @@ def make_custom_env(
 
 
 # most of this function is redundant as it is copied from sample.enjoy.enjoy
-def create_dataset(cfg: Config, dataset_size: int = 100_000, split: str = "train") -> None:
+def create_dataset(cfg: Config) -> None:
     cfg = load_from_checkpoint(cfg)
     eval_env_frameskip: int = cfg.env_frameskip if cfg.eval_env_frameskip is None else cfg.eval_env_frameskip
     assert (
@@ -122,6 +68,7 @@ def create_dataset(cfg: Config, dataset_size: int = 100_000, split: str = "train
     rnn_states = torch.zeros([env.num_agents, get_rnn_size(cfg)], dtype=torch.float32, device=device)
 
     # Run the environment
+    dataset_size = 1_600_000 + 160_000
     progress_bar = tqdm(total=dataset_size)
     num_timesteps = 0
     with torch.no_grad():
@@ -142,7 +89,8 @@ def create_dataset(cfg: Config, dataset_size: int = 100_000, split: str = "train
 
             # Actions shape should be [num_agents, num_actions] even if it's [1, 1]
             actions = preprocess_actions(env_info, actions)
-
+            # Clamp actions to be in the range of the action space
+            actions = np.clip(actions, env.action_space.low, env.action_space.high)
             rnn_states = policy_outputs["new_rnn_states"]
             dataset["continuous_observations"][-1].append(observations["obs"].cpu().numpy()[0])
             dataset["continuous_actions"][-1].append(actions[0])
@@ -166,17 +114,21 @@ def create_dataset(cfg: Config, dataset_size: int = 100_000, split: str = "train
 
     repo_path = f"datasets/{cfg.experiment[:-3]}"
     os.makedirs(repo_path, exist_ok=True)
-    file = f"{repo_path}/{split}"
-    np.savez_compressed(f"{file}.npz", **dataset)
+
+    _dataset = {key: value[:16_000] for key, value in dataset.items()}
+    file = f"{repo_path}/train"
+    np.savez_compressed(f"{file}.npz", **_dataset)
+
+    _dataset = {key: value[16_000:] for key, value in dataset.items()}
+    file = f"{repo_path}/test"
+    np.savez_compressed(f"{file}.npz", **_dataset)
 
 
 def main() -> int:
-    for env_name in ENV_NAMES:
-        register_env(env_name, make_custom_env)
     parser, _ = parse_sf_args(argv=None, evaluation=True)
     cfg = parse_full_cfg(parser)
-    status = create_dataset(cfg, dataset_size=1_600_000, split="train")
-    status = create_dataset(cfg, dataset_size=160_000, split="test")
+    register_env(cfg.env, make_custom_env)
+    status = create_dataset(cfg)
     return status
