+++ data/envs/metaworld/train.py
@@ -2,65 +2,11 @@ import argparse
 import sys
 from typing import Dict, Optional
 
-import gym
+import gymnasium as gym
 import metaworld  # noqa: F401
 from sample_factory.cfg.arguments import parse_full_cfg, parse_sf_args
 from sample_factory.envs.env_utils import register_env
 from sample_factory.train import run_rl
-
-
-ENV_NAMES = [
-    "assembly-v2",
-    "basketball-v2",
-    "bin-picking-v2",
-    "box-close-v2",
-    "button-press-topdown-v2",
-    "button-press-topdown-wall-v2",
-    "button-press-v2",
-    "button-press-wall-v2",
-    "coffee-button-v2",
-    "coffee-pull-v2",
-    "coffee-push-v2",
-    "dial-turn-v2",
-    "disassemble-v2",
-    "door-close-v2",
-    "door-lock-v2",
-    "door-open-v2",
-    "door-unlock-v2",
-    "drawer-close-v2",
-    "drawer-open-v2",
-    "faucet-close-v2",
-    "faucet-open-v2",
-    "hammer-v2",
-    "hand-insert-v2",
-    "handle-press-side-v2",
-    "handle-press-v2",
-    "handle-pull-side-v2",
-    "handle-pull-v2",
-    "lever-pull-v2",
-    "peg-insert-side-v2",
-    "peg-unplug-side-v2",
-    "pick-out-of-hole-v2",
-    "pick-place-v2",
-    "pick-place-wall-v2",
-    "plate-slide-back-side-v2",
-    "plate-slide-back-v2",
-    "plate-slide-side-v2",
-    "plate-slide-v2",
-    "push-back-v2",
-    "push-v2",
-    "push-wall-v2",
-    "reach-v2",
-    "reach-wall-v2",
-    "shelf-place-v2",
-    "soccer-v2",
-    "stick-pull-v2",
-    "stick-push-v2",
-    "sweep-into-v2",
-    "sweep-v2",
-    "window-close-v2",
-    "window-open-v2",
-]
 
 
 def make_custom_env(
@@ -79,7 +25,7 @@ def override_defaults(parser: argparse.ArgumentParser) -> argparse.ArgumentParse
         num_workers=8,
         num_envs_per_worker=8,
         worker_num_splits=2,
-        train_for_env_steps=100_000_000,
+        train_for_env_steps=10_000_000,
         encoder_mlp_layers=[64, 64],
         env_frameskip=1,
         nonlinearity="tanh",
@@ -116,11 +62,10 @@ def override_defaults(parser: argparse.ArgumentParser) -> argparse.ArgumentParse
 
 
 def main() -> int:
-    for env_name in ENV_NAMES:
-        register_env(env_name, make_custom_env)
     parser, _ = parse_sf_args(argv=None, evaluation=False)
     parser = override_defaults(parser)
     cfg = parse_full_cfg(parser)
+    register_env(cfg.env, make_custom_env)
     status = run_rl(cfg)
     return status
